requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
flask-restx==1.3.0
Flask-SocketIO==5.3.7
Jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.2